+++ actix-web/src/lib.rs
@@ -155,6 +155,7 @@ codegen_reexport!(delete);
 codegen_reexport!(trace);
 codegen_reexport!(connect);
 codegen_reexport!(options);
+codegen_reexport!(query);
 codegen_reexport!(scope);
 
 pub(crate) type BoxError = Box<dyn std::error::Error>;