+++ actix-web/src/test/test_request.rs
@@ -111,6 +111,12 @@ impl TestRequest {
         TestRequest::default().method(Method::PATCH)
     }
 
+    /// Constructs test request with QUERY method.
+    // TODO: use `Method::QUERY` once the `http` dependency is bumped (see #3384).
+    pub fn query() -> TestRequest {
+        TestRequest::default().method(Method::from_bytes(b"QUERY").unwrap())
+    }
+
     /// Constructs test request with DELETE method.
     pub fn delete() -> TestRequest {
         TestRequest::default().method(Method::DELETE)