+++ actix-web-codegen/CHANGES.md
@@ -2,8 +2,11 @@
 
 ## Unreleased
 
+- Add `#[query]` macro for routing the HTTP `QUERY` method (RFC 10008). [#4140]
 - Minimum supported Rust version (MSRV) is now 1.88.
 - Update `syn` dependency to `0.3`.
+
+[#4140]: https://github.com/actix/actix-web/pull/4140
 
 ## 4.3.0